moved connect.go to models package
