+++ internal/models/connect_test.go
@@ -1,4 +1,4 @@
-package sql
+package models
 
 import (
 	"strconv"
