+++ internal/reddit/grab.go
@@ -29,7 +29,7 @@ var (
 
 // GrabSaved reads all cached posts on the Reddit account.
 // This can be used to mass refresh an entire SQL database.
-func GrabSaved(postsTable, commentsTable *models.Table[models.Row[models.id, text]], key *models.Key) {
+func GrabSaved(postsTable, commentsTable *models.Table[models.Row[models.Id, models.Text]], key *models.Key) {
 
 	var mySavedPosts []*reddit.Post
 	var mySavedComments []*reddit.Comment
